+++ src/__tests__/pages/article.test.tsx
@@ -6,6 +6,7 @@ import axios from 'axios'
 import { useRouter } from 'next/router'
 import type { GetServerSidePropsContext } from 'next'
 
+import { DateTime } from 'luxon'
 import { renderWithAuth } from '../../testHelpers'
 import { client } from '../../lib/keystoneClient'
 
@@ -13,6 +14,8 @@ import { cmsOrbitBlogArticle as mockOrbitBlogArticle } from '../../__fixtures__/
 import { cmsInternalNewsArticle } from '../../__fixtures__/data/cmsInternalNewsArticle'
 
 import SingleArticlePage, { getServerSideProps } from 'pages/articles/[article]'
+import { testUser1, cmsUser } from '__fixtures__/authUsers'
+import { getSession } from 'lib/session'
 
 jest.mock('../../lib/keystoneClient', () => ({
   client: {
@@ -55,6 +58,15 @@ mockedUseRouter.mockReturnValue({
   replace: mockReplace,
 })
 
+jest.mock('lib/session', () => ({
+  getSession: jest.fn(),
+}))
+
+const mockedGetSession = getSession as jest.Mock
+mockedGetSession.mockImplementationOnce(() =>
+  Promise.resolve({ passport: { user: testUser1 } })
+)
+
 describe('Single article getServerSideProps', () => {
   const testContext = {
     query: { article: 'test-article-slug' },
@@ -69,10 +81,56 @@ describe('Single article getServerSideProps', () => {
     })
   })
 
+  it('returns found if the query returns an unpublished article but user is a cms user', async () => {
+    const draftArticle = { ...mockOrbitBlogArticle, status: 'Draft' }
+    mockedKeystoneClient.query.mockResolvedValueOnce({
+      data: {
+        article: draftArticle,
+      },
+      loading: false,
+      errors: [],
+      networkStatus: 7,
+    })
+
+    mockedGetSession.mockImplementationOnce(() =>
+      Promise.resolve({ passport: { user: cmsUser } })
+    )
+
+    const response = await getServerSideProps(testContext)
+
+    expect(response).toEqual({
+      props: {
+        article: draftArticle,
+      },
+    })
+  })
+
   it('returns not found if the query returns an unpublished article', async () => {
     mockedKeystoneClient.query.mockResolvedValueOnce({
       data: {
         article: { ...mockOrbitBlogArticle, status: 'Draft' },
+      },
+      loading: false,
+      errors: [],
+      networkStatus: 7,
+    })
+
+    const response = await getServerSideProps(testContext)
+
+    expect(response).toEqual({
+      notFound: true,
+    })
+  })
+
+  it('returns not found if the query returns an article published in the future', async () => {
+    const futureDate = DateTime.now().plus({ weeks: 2 })
+    mockedKeystoneClient.query.mockResolvedValueOnce({
+      data: {
+        article: {
+          ...mockOrbitBlogArticle,
+          status: 'Published',
+          publishedDate: futureDate.toISO(),
+        },
       },
       loading: false,
       errors: [],