+++ src/components/SingleArticle/SingleArticle.test.tsx
@@ -6,14 +6,16 @@ import { render, screen } from '@testing-library/react'
 import React from 'react'
 
 import { SingleArticle } from './SingleArticle'
+import type { ArticleRecord } from 'types'
 
-const testArticle = {
+const testArticle: ArticleRecord = {
   id: 'testArticleId123',
   slug: 'test-article',
   title:
     'Version 2.8.5 released! Includes MVP search experience and a way to filter the news.',
   category: 'InternalNews',
   publishedDate: '2022-05-17T13:44:39.796Z',
+  status: 'Published',
   hero: {
     url: 'http://cms.example.com/images/image.png',
   },
@@ -163,6 +165,8 @@ describe('SingleArticle component', () => {
   test('renders the article', () => {
     render(<SingleArticle article={testArticle} />)
 
+    const banner = screen.queryByText('Unpublished Article Preview')
+    expect(banner).toBeNull()
     expect(screen.getByText('May 17, 2022')).toBeInTheDocument()
     expect(screen.getByRole('heading', { level: 2 })).toHaveTextContent(
       testArticle.title
@@ -189,5 +193,14 @@ describe('SingleArticle component', () => {
       name: 'article hero graphic',
     })
     expect(img).not.toBeInTheDocument()
+  })
+
+  test('renders the unpublished article preview banner if article is not published', () => {
+    const unpublished: ArticleRecord = { ...testArticle, status: 'Draft' }
+    render(<SingleArticle article={unpublished} />)
+
+    const banner = screen.getByText('Unpublished Article Preview')
+    expect(banner).toBeVisible()
+    expect(banner).toHaveClass('previewBanner')
   })
 })