+++ src/operations/cms/queries/getInternalNewsCarouselArticles.ts
@@ -1,10 +1,11 @@
 import { gql } from '@apollo/client'
 
 export const GET_INTERNAL_NEWS_CAROUSEL_ARTICLES = gql`
-  query GetInternalNewsCarouselArticles {
+  query GetInternalNewsCarouselArticles($publishedDate: DateTime) {
     articles(
       where: {
         status: { equals: Published }
+        publishedDate: { lte: $publishedDate }
         category: { equals: InternalNews }
       }
       orderBy: [{ publishedDate: desc }]